backend/classifier-api/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
sqlalchemy
python-dotenv
psycopg2-binary